requirements.txt
ADMCode==0.5.2
alabaster==0.7.10
anaconda-client==1.6.9
anaconda-navigator==1.7.0
anaconda-project==0.8.2
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
arabic-reshaper==2.0.14
arrow==0.12.1
asn1crypto==0.24.0
astroid==1.6.1
astropy==2.0.3
astunparse==1.6.2
atom==0.4.1
attrs==19.1.0
Babel==2.5.3
backports.lzma==0.0.13
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
binaryornot==0.4.4
bitarray==0.8.1
bkcharts==0.2
black==19.3b0
blaze==0.11.3
bleach==2.1.2
bokeh==0.12.13
boto==2.48.0
Bottleneck==1.2.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
conda==4.6.14
conda-build==3.4.1
conda-verify==2.0.0
configobj==5.0.6
configparser==3.7.4
contextlib2==0.5.5
cookiecutter==1.6.0
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.7
cytoolz==0.9.0
dabest==0.1.5
dask==0.16.1
datashape==0.5.4
decorator==4.2.1
distributed==1.20.2
docutils==0.14
entrypoints==0.2.3
esprima==4.0.1
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.13
Flask==0.12.2
Flask-Cors==3.0.3
freetype-py==2.1.0.post1
future==0.16.0
gevent==1.4.0
gitdb2==2.0.5
GitPython==2.1.11
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.15
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipykernel==5.1.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.11.1
Jinja2==2.10
jinja2-time==0.2.0
joblib==0.13.2
joypy==0.1.10
json-tricks==3.13.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.31.5
jupyterlab-launcher==0.10.2
jupyterthemes==0.19.1
kabuki==0.6.2
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
lesscpy==0.13.0
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.3
MarkupSafe==1.0
matplotlib==3.0.3
mccabe==0.6.1
mgcpy==0.3.0
mistune==0.8.3
mkl-fft==1.0.12
mkl-random==1.0.2
more-itertools==4.3.0
moviepy==0.2.3.5
mpmath==1.0.0
msgpack==0.6.1
msgpack-numpy==0.4.4.2
msgpack-python==0.5.6
multipledispatch==0.4.9
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.3
numpydoc==0.7.0
odo==0.5.1
olefile==0.45.1
opencv-python==4.1.0.25
openpyxl==2.6.2
packaging==16.8
pandas==0.24.2
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.1
pep8==1.7.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==6.0.0
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
poyo==0.4.1
prometheus-client==0.6.0
prompt-toolkit==1.0.15
psutil==5.6.2
psychopy-ext==0.6.0.4
ptyprocess==0.5.2
py==1.5.2
pybind11==2.2.4
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==1.6.0
pygame==1.9.6
pyglet==1.3.2
Pygments==2.2.0
pylint==1.8.2
pymc==2.3.6
pyobjc-core==5.2
pyobjc-framework-Cocoa==5.2
pyobjc-framework-Quartz==5.2
pyodbc==4.0.22
PyOpenGL==3.1.1a1
pyOpenSSL==17.5.0
pyosf==1.0.5
pyparallel==0.2.2
pyparsing==2.2.0
pyserial==3.4
PySocks==1.6.7
PySoundCard==0.5.2
pytest==3.3.2
python-bidi==0.4.0
python-dateutil==2.6.1
python-gitlab==1.8.0
pytz==2017.3
PyWavelets==0.5.2
PyYAML==5.1
pyzmq==17.1.2
QtAwesome==0.5.7
qtconsole==4.3.1
QtPy==1.7.1
requests==2.21.0
rerf==2.0.2
rope==0.10.7
rpy2==2.9.4
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.21.1
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.4.2
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smmap2==2.0.5
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
sounddevice==0.3.13
SoundFile==0.10.1
spark-parser==1.8.7
Sphinx==1.6.6
sphinxcontrib-websupport==1.0.1
spyder==3.3.4
spyder-kernels==0.4.4
SQLAlchemy==1.2.1
statsmodels==0.9.0
sympy==1.1.1
tables==3.5.1
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
toml==0.10.0
toolz==0.9.0
tornado==4.5.3
tqdm==4.32.1
traitlets==4.3.2
typing==3.6.2
uncompyle6==3.3.4
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whichcraft==0.4.1
widgetsnbextension==3.1.0
wrapt==1.10.11
wurlitzer==1.0.2
wxPython==4.0.4
xdis==4.0.2
xlrd==1.2.0
XlsxWriter==1.0.2
xlwings==0.11.5
xlwt==1.2.0
zict==0.1.3
zmq==0.0.0
